[Refactor] UIManager에서 LobbyUIManager 파일 분리
